METL-694: updated StronglyTypedJsonActor to require a list of JValues as the response instead of a single JValue, so that functions can be called in JS which take multiple arguments.  Reworked all the calls in comet\MeTL.scala to suit.  This will enable sending the messages directly back into MeTLBus, instead of bouncing off global methods which have to be defined on JS.

+++ src/main/scala/com/metl/model/metl2011/MeTL2011XmlSerializer.scala
@@ -98,7 +98,7 @@ class MeTL2011XmlSerializer(config:ServerConfiguration,cacheImages:Boolean = fal
       val highlight = getColorByName(bl,"highlight")
       SubmissionBlacklistedPerson(username,highlight)
     }).toList
-    MeTLSubmission(config,m.author,m.timestamp,title,c.slide.toInt,url,imageBytes,blacklist,c.target,c.privacy,c.identity,m.audiences)
+    MeTLSubmission(config,m.author,m.timestamp,title,c.slide,url,imageBytes,blacklist,c.target,c.privacy,c.identity,m.audiences)
   })
   override def toMeTLQuiz(input:NodeSeq):MeTLQuiz = Stopwatch.time("MeTL2011XmlSerializer.toMeTLQuiz",{
 		trace("quiz attempted: %s".format(input))
